Всё готово надеюьсб

+++ controllers/commentController.js
@@ -3,8 +3,14 @@ import CommentModule from '../models/comments.js';
 
 export const createComment = async (req, res) => {
 
-    const {user_id , text} = req.body;
+    const {text} = req.body;
+    const user_id = req.userId.id;
     const post_id = req.params.id
+
+    if (!user_id || !post_id){
+        return res.status(404).json('User of Post not found');
+    }
+
     try{
 
         const comment = await CommentModule.create({user_id , post_id, comment: text})
@@ -21,8 +27,13 @@ export const createComment = async (req, res) => {
 
 export const deleteComment = async (req, res) => {
 
-    const {user_id} = req.body;
+    // const {user_id} = req.body;
+    const user_id = req.userId.id;
     const post_id = req.params.id;
+
+    if (!user_id || !post_id){
+        return res.status(404).json('User of Post not found');
+    }
 
     try{
         const comment = await CommentModule.deleteOne({
@@ -44,8 +55,13 @@ export const deleteComment = async (req, res) => {
 
 export const updateComment = async (req, res) => {
 
-    const {user_id, text} = req.body;
+    const {text} = req.body;
+    const user_id = req.userId.id
     const post_id = req.params.id;
+
+    if (!user_id || !post_id){
+        return res.status(404).json('User of Post not found');
+    }
 
     try{
         const adlike = await CommentModule.updateOne({
@@ -68,6 +84,10 @@ export const updateComment = async (req, res) => {
 export const getComments = async (req, res) => {
 
     const post_id = req.params.id;
+
+    if  (!post_id){
+        return res.status(404).json(' Post not found');
+    }
 
     try{
         const commetns = await CommentModule.find({ post_id : post_id });
